chore(tests): consolidate the orphaned tests/ dir into the collected test/ tree
The root tests/ directory predates setup.cfg's testpaths pin (testpaths =
test src/kiro_crew/apps/builtins) and is never collected by pytest or CI —
yet its files kept receiving edits that never ran. Running it: 144 passed,
18 failed on stale-API drift, 1 import error.

Moves (now actually run in CI):
- test_changelog_parse, test_conductor_skill, test_gatewayd_diag,
  test_select_crew, test_session_workspace, test_subagent_startup_watchdog
  — clean, moved as-is (81 tests).
- test_config_roundtrip — to_dict-completeness contract fixed to exclude
  underscore-private loader state; 3 tests for the removed SecretaryConfig
  deleted (5 tests).
- test_context_management — 5 tests importing plan-memory functions that
  never existed in this repo's history deleted (34 tests).
- test_temporary_chat — save-slot tests re-pointed at
  chat_persistence.slot_history_key (the real binding; the old chat.py
  patch target missed, letting mocks hit json.dump); modifier test updated
  to the _apply_privacy_mode call shape (set_slack_link gets the bare
  reply_ts); redundant bounded-eviction test deleted (canonical
  test_messaging_privacy_mode covers PRIVACY_LRU_MAX eviction) (18 tests).
- tests/test_context.py — merged into test/test_context.py (unique
  WIDGET_BLOCK/MAX_SUBAGENTS placeholder coverage; the stale 300-char
  budget re-scoped per-section since the Artifacts section was added to
  the same template after the test was written).

Deleted (dead):
- test_agent_metadata.py — its module was removed in #1367.
- test_subagent_validate.py — redundant with test_subagent_coverage.py's
  real-module coverage; its sys.modules stubs had drifted unmaintainably.

Deliberately NOT moved: tests/workflows/malicious/ — the B9 adversarial
corpus, placed outside testpaths precisely so pytest never collects
hostile workflow scripts as tests; test_workflows_malicious.py loads it
by that path (B9 gate re-verified, 15 pass).

no linked issue: cleanup surfaced during portfolio work; no bug filed.

153 tests green across the moved suites + B9; 98 in the merged
test_context.py (2 pre-existing local-venv failures reproduce on pristine
main and are unrelated).
