+++ test/test_context.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from types import SimpleNamespace
 from unittest.mock import patch
 
 import pytest
@@ -261,9 +262,7 @@ class TestContextBuilder:
         """With the flag set on a continuing session, the index comes back
         wrapped in the marker so the model can still discover skills."""
         builder = self._reinject_builder(tmp_path)
-        msg, _ = builder.build_message(
-            "carry on", is_new_session=False, needs_reinjection=True
-        )
+        msg, _ = builder.build_message("carry on", is_new_session=False, needs_reinjection=True)
         assert "[REINJECTED AFTER COMPACTION" in msg
         assert "[END REINJECTED]" in msg
         assert "widget-maker" in msg, "the re-injected block must carry the skill index"
@@ -278,9 +277,7 @@ class TestContextBuilder:
         """A new session already gets the index from the session context;
         re-injecting would duplicate it in the same prompt."""
         builder = self._reinject_builder(tmp_path)
-        msg, _ = builder.build_message(
-            "first turn", is_new_session=True, needs_reinjection=True
-        )
+        msg, _ = builder.build_message("first turn", is_new_session=True, needs_reinjection=True)
         assert "[REINJECTED AFTER COMPACTION" not in msg
 
     def test_no_reinjection_for_an_unmapped_custom_agent(self, tmp_path):
@@ -393,9 +390,7 @@ class TestContextBuilder:
             memory=MemoryStore(workspace=tmp_path / "ws"),
             skills=SkillsLoader(skills_path=tmp_path / "skills", install_builtins=False),
         )
-        msg, _ = builder.build_message(
-            "hello", is_new_session=False, folder_path="Backend › 0812"
-        )
+        msg, _ = builder.build_message("hello", is_new_session=False, folder_path="Backend › 0812")
         assert "[FOLDER]" in msg
         assert "Backend › 0812" in msg
 
@@ -1353,3 +1348,143 @@ class TestMemoryGetContextQueryWiring:
         )
         builder.build_message("find my tokyo notes", True, "s2")
         assert seen == ["find my tokyo notes"]
+
+
+# ── {{WIDGET_BLOCK}} / {{MAX_SUBAGENTS}} placeholder resolution ──────────
+# (Merged from the orphaned tests/test_context.py during the tests/ dir
+# consolidation — see the docstring history in that file's git log.)
+
+
+def _resolve(prompt: str, session_key: str, density: str = "more") -> str:
+    """Call the private template resolver with a canned config."""
+    from kiro_crew.context import ContextBuilder
+
+    fake_cfg = SimpleNamespace(dashboard=SimpleNamespace(widget_density=density))
+    with patch("kiro_crew.context.KiroCrewConfig.load", return_value=fake_cfg):
+        return ContextBuilder._resolve_prompt_templates(prompt, session_key)
+
+
+class TestWidgetBlockPlaceholder:
+    """`{{WIDGET_BLOCK}}` expands to a short skill pointer on dashboard; empty elsewhere."""
+
+    def test_non_dashboard_strips_placeholder(self):
+        # Slack / CLI / channel sessions cannot render widgets; the placeholder
+        # MUST expand to empty so the prompt stays lean.
+        prompt = "prefix {{WIDGET_BLOCK}} suffix"
+        for key in ("slack:C123:123.456", "cli:local", "channel:Cabc", ""):
+            result = _resolve(prompt, key)
+            assert "{{WIDGET_BLOCK}}" not in result
+            assert "mcwidget" not in result.lower()
+            assert result == "prefix  suffix"
+
+    def test_dashboard_more_density_emits_pointer(self):
+        # The `more` branch should encourage widgets and point at the skill.
+        result = _resolve("{{WIDGET_BLOCK}}", "dashboard:abc", density="more")
+        assert "## Inline Widgets" in result
+        assert "<mcwidget" in result
+        assert "`widgets` skill" in result
+
+    def test_dashboard_less_density_emits_pointer(self):
+        # The `less` branch should discourage widgets but still point at the skill.
+        result = _resolve("{{WIDGET_BLOCK}}", "dashboard:abc", density="less")
+        assert "## Inline Widgets" in result
+        assert "<mcwidget" in result
+        assert "`widgets` skill" in result
+        assert "prefer" in result.lower()
+
+    def test_pointer_does_not_restate_skill_content(self):
+        # The main-prompt pointer must NOT inline the theme-variable table,
+        # the full rules, or per-Tailwind-class guidance. That lives in the
+        # bundled skill. Regression guard against reintroducing the bloat.
+        for density in ("more", "less"):
+            result = _resolve("{{WIDGET_BLOCK}}", "dashboard:abc", density=density)
+            assert "var(--bg)" not in result, f"theme var leaked into {density} pointer"
+            assert "var(--card)" not in result
+            assert "Chart.js" not in result
+            assert "bg-[var(" not in result
+
+    def test_pointer_is_short(self):
+        # Budget guard against re-inlining bulk guidance. The block now carries
+        # TWO pointer sections (Inline Widgets + Artifacts, the latter added
+        # after this test was written), so budget each section independently:
+        # the regression this guards (inlining the ~800-char theme-variable
+        # table or the full artifacts workflow into a pointer) trips its
+        # section's own cap regardless of how small the other section is.
+        for density in ("more", "less"):
+            result = _resolve("{{WIDGET_BLOCK}}", "dashboard:abc", density=density)
+            widgets_section, _, artifacts_section = result.partition("## Artifacts")
+            assert (
+                len(widgets_section) < 400
+            ), f"{density} widgets section too long: {len(widgets_section)} chars"
+            assert (
+                len(artifacts_section) < 500
+            ), f"{density} artifacts section too long: {len(artifacts_section)} chars"
+
+    def test_dashboard_underscore_key_also_matches(self):
+        # Some dashboard sessions use `dashboard_<slot>` instead of `dashboard:<slot>`.
+        result = _resolve("{{WIDGET_BLOCK}}", "dashboard_slot1", density="more")
+        assert "<mcwidget" in result
+
+    def test_density_default_when_config_missing(self):
+        # If the dashboard config omits widget_density entirely, fall back to "more".
+        from kiro_crew.context import ContextBuilder
+
+        fake_cfg = SimpleNamespace(dashboard=SimpleNamespace())
+        with patch("kiro_crew.context.KiroCrewConfig.load", return_value=fake_cfg):
+            result = ContextBuilder._resolve_prompt_templates("{{WIDGET_BLOCK}}", "dashboard:x")
+        # The `more` branch fires (skill pointer + encouraging wording) and the
+        # `less` branch does not. Assert structurally, not on specific prose
+        # tokens — the wording may evolve.
+        assert "`widgets` skill" in result, "skill pointer missing"
+        assert "prefer" not in result.lower(), "less-branch wording leaked"
+
+
+class TestMaxSubagentsPlaceholder:
+    """`{{MAX_SUBAGENTS}}` expands to the live resolved concurrent cap on every transport."""
+
+    @staticmethod
+    def _resolve_cap(prompt, session_key, *, cap=None, raises=False):
+        from kiro_crew.context import ContextBuilder
+
+        fake_cfg = SimpleNamespace(dashboard=SimpleNamespace(widget_density="more"))
+        if raises:
+            sub = patch(
+                "kiro_crew.subagent.resolve_max_subagents",
+                side_effect=RuntimeError("boom"),
+            )
+        else:
+            sub = patch("kiro_crew.subagent.resolve_max_subagents", return_value=cap)
+        with patch("kiro_crew.context.KiroCrewConfig.load", return_value=fake_cfg), sub:
+            return ContextBuilder._resolve_prompt_templates(prompt, session_key)
+
+    def test_token_replaced_with_live_cap_on_every_transport(self):
+        # The cap must reach dashboard, Slack, CLI, and empty-key sessions alike —
+        # delegation guidance is transport-agnostic.
+        for key in ("dashboard:abc", "slack:C1:1.2", "cli:local", ""):
+            result = self._resolve_cap("up to {{MAX_SUBAGENTS}} agents", key, cap=12)
+            assert "{{MAX_SUBAGENTS}}" not in result
+            assert "up to 12 agents" in result
+
+    def test_zero_cap_falls_back_to_several(self):
+        # cap==0 (auto-size failed / unreadable host) keeps the sentence grammatical.
+        result = self._resolve_cap("up to {{MAX_SUBAGENTS}} agents", "slack:C1:1.2", cap=0)
+        assert "up to several agents" in result
+
+    def test_resolver_error_falls_back_to_several(self):
+        # A raising resolver must never break prompt assembly.
+        result = self._resolve_cap("up to {{MAX_SUBAGENTS}} agents", "dashboard:abc", raises=True)
+        assert "up to several agents" in result
+
+    def test_absent_token_skips_resolver(self):
+        # No token → the (heavier) sub-agent resolver is never invoked.
+        from kiro_crew.context import ContextBuilder
+
+        fake_cfg = SimpleNamespace(dashboard=SimpleNamespace(widget_density="more"))
+        with (
+            patch("kiro_crew.context.KiroCrewConfig.load", return_value=fake_cfg),
+            patch("kiro_crew.subagent.resolve_max_subagents") as resolver,
+        ):
+            ContextBuilder._resolve_prompt_templates(
+                "no token here {{WIDGET_BLOCK}}", "dashboard:abc"
+            )
+        resolver.assert_not_called()
