+++ test/test_gatewayd_diag.py
